TST: collect Period tests
