+++ pandas/tests/tseries/frequencies/test_freq_code.py
@@ -3,11 +3,8 @@ import pytest
 
 from pandas._libs.tslibs import (
     Period,
-    Resolution,
     to_offset,
 )
-
-import pandas._testing as tm
 
 
 @pytest.mark.parametrize(
@@ -21,27 +18,6 @@ def test_get_to_timestamp_base(freqstr, exp_freqstr):
 
     result_code = per._dtype._get_to_timestamp_base()
     assert result_code == exp_code
-
-
-@pytest.mark.parametrize(
-    "freqstr,expected",
-    [
-        ("Y", "year"),
-        ("Q", "quarter"),
-        ("M", "month"),
-        ("D", "day"),
-        ("h", "hour"),
-        ("min", "minute"),
-        ("s", "second"),
-        ("ms", "millisecond"),
-        ("us", "microsecond"),
-        ("ns", "nanosecond"),
-    ],
-)
-def test_get_attrname_from_abbrev(freqstr, expected):
-    reso = Resolution.get_reso_from_freqstr(freqstr)
-    assert reso.attr_abbrev == freqstr
-    assert reso.attrname == expected
 
 
 @pytest.mark.parametrize(
@@ -91,12 +67,3 @@ def test_compatibility(freqstr, expected):
     ts_np = np.datetime64("2021-01-01T08:00:00.00")
     do = to_offset(freqstr)
     assert ts_np + do == np.datetime64(expected)
-
-
-@pytest.mark.parametrize("freq", ["A", "H", "T", "S", "L", "U", "N"])
-def test_units_A_H_T_S_L_U_N_deprecated_from_attrname_to_abbrevs(freq):
-    # GH#52536
-    msg = f"'{freq}' is deprecated and will be removed in a future version."
-
-    with tm.assert_produces_warning(FutureWarning, match=msg):
-        Resolution.get_reso_from_freqstr(freq)
